requirements.txt
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.13.2
chardet==4.0.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
decorator==4.4.1
EasyProcess==0.2.9
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
lxml==4.4.2
parsel==1.5.2
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
Protego==0.1.16
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.5.2
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
queuelib==1.5.0
requests==2.25.1
Scrapy==1.8.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.2
traitlets==4.3.3
Twisted==19.10.0
urllib3==1.25.8
w3lib==1.21.0
wcwidth==0.1.8
zope.interface==4.7.1
